telethon-listener/requirements.txt
aiohttp==3.9.3
telethon==1.34.0
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0 